+++ packages/extractor-vue-script/build.config.ts
@@ -6,6 +6,9 @@ export default defineBuildConfig({
   ],
   declaration: true,
   clean: true,
+  rollup: {
+    emitCJS: true,
+  },
   externals: [
     '@unocss/core',
   ],